Created package specifically for scraping, created a separate class for navigating home page called navigator.
Ran into problem using b-soup with dynamic web page. Implemented selenium to get HTML as it is when loaded. Selenium problems are:
1. speed
2. Opening a browser when run - solution; consider headless web driver

Should also consider creating new django application for scraping and storing data. Its to slow doing it every pageload. It can be in the same project but a different django app.